閾値の速度に関するUIを変更

+++ app/src/main/java/com/example/getofftopause/MediaControlService.java
@@ -95,7 +95,7 @@ public class MediaControlService extends Service implements AudioManager.OnAudio
                 return;
             }
 
-            float speedThresholdKph = sharedPreferences.getFloat(getString(R.string.speed_threshold_key), 8);
+            int speedThresholdKph = sharedPreferences.getInt(getString(R.string.speed_threshold_key), 8);
 
             float lastSpeedMps = location.getSpeed();
             float lastSpeedKph = 3.6f * lastSpeedMps;
